+++ example/mlx/buildupRT/src/product.c
@@ -1,6 +1,6 @@
 #include "vec3.h"
 
-float	ft_vec3_dot(t_vec3 *u, t_vec3 *v)
+double	ft_vec3_dot(t_vec3 *u, t_vec3 *v)
 {
 	return (u->x * v->x + u->y * v->y + u->z * v->z);
 }
